backend/requirements/dev.txt
-r base.txt
pytest-django==4.7.0
pytest-cov==4.1.0